Add Event Get Api & NotFound

+++ src/main/java/me/study/springrestapi/events/EventRepository.java
@@ -2,5 +2,5 @@ package me.study.springrestapi.events;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
 }
